Renamed e2scontroller to controller
